requirements.txt
# Web scraping and API interaction
requests==2.31.0
beautifulsoup4==4.12.2

# Data processing
pandas==2.1.0
numpy==1.25.2
tabulate==0.9.0

# Dashboard and visualization
dash==2.13.0
plotly==5.17.0
dash-bootstrap-components==1.5.0

# Testing
pytest==7.4.0

# Utilities
python-dotenv==1.0.0 